Added auth and responses
Added authorization to the endpoints, and now return responses using the new ResponseMessage class. Further, endpoints that are not going to be used have been commented out and return 403.

+++ backend/user/src/main/java/edu/oswego/rest/controller/user/UserAPI.java
@@ -2,22 +2,33 @@ package edu.oswego.rest.controller.user;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.google.gson.Gson;
+import com.ibm.websphere.security.jwt.InvalidConsumerException;
+import com.ibm.websphere.security.jwt.InvalidTokenException;
+import com.ibm.websphere.security.jwt.JwtConsumer;
 import edu.oswego.util.objects.User;
+import edu.oswego.util.objects.authObject;
 import edu.oswego.util.service.IUserService;
 import edu.oswego.util.service.impl.UserService;
+import edu.oswego.util.utility.ResponseMessage;
 // Json-B
 import javax.json.bind.Jsonb;
 import javax.json.bind.JsonbBuilder;
 
 // JAX-RS
 import javax.ws.rs.*;
+import javax.ws.rs.core.Response;
 import java.util.List;
 
 @Path("/user")
 public class UserAPI {
     private static final long serialVersionUID = 1L;
     private IUserService userService;
-    private Jsonb jsonb = JsonbBuilder.create();
+    private Jsonb jsonb = JsonbBuilder.create();private final authObject a = new authObject();
+    private final String unauth = "Unauthorized: You mus log in to view this page.";
+    private final String forbid = "Forbidden: You do not have access to this page.";
+    private final String errorm = "An error occurred: ";
+    private final String badnum = "The ID number provided was not formatted properly.";
+
 
 
     public UserAPI() {
@@ -25,8 +36,8 @@ public class UserAPI {
     }
 
     @GET
-    public String getAllUsers(){
-        //TODO This method needs to ensure authentication
+    public Response getAllUsers(){
+        /*
         try {
             List<User> listOfUsers = userService.findAll();
             String res = jsonb.toJson(listOfUsers);
@@ -35,15 +46,15 @@ public class UserAPI {
         } catch (NumberFormatException ne){
             System.out.println("There are not any users in the database.");
         }
-        return null;
+         */
+        return new ResponseMessage().sendMessage(forbid, 403);
     }
 
     @GET
     @Path("/{userId}")
-    public String getSpecificUser(@PathParam("userId") String _userId){
-        //TODO This method needs to ensure authentication
+    public Response getSpecificUser(@PathParam("userId") String _userId){
+        /*
         try {
-
             int userId = Integer.parseInt(_userId);
             User user = userService.findOne(userId);
             String res = jsonb.toJson(user);
@@ -52,39 +63,62 @@ public class UserAPI {
         } catch (NumberFormatException ne){
             System.out.println("User ID provided was not formatted properly.");
         }
-        return null;
+        */
+        return new ResponseMessage().sendMessage(forbid, 403);
     }
     @POST
-    public String postUser(String payload) throws JsonProcessingException {
+    public Response postUser(String payload) {
+        /*
         User user = jsonb.fromJson(payload, User.class);
         user = userService.save(user);
         String res = jsonb.toJson(user);
         return res;
+         */
+        return new ResponseMessage().sendMessage(forbid, 403);
     }
 
     @PUT
-    public String updateUser(String payload) throws JsonProcessingException {
+    public Response updateUser(String payload) {
+        /*
         User user = jsonb.fromJson(payload, User.class);
         user = userService.update(user);
         String res = jsonb.toJson(user);
-        return res;
+         */
+        return new ResponseMessage().sendMessage(forbid, 403);
     }
 
+    /**
+     * This method only allows professors to delete student user objects.
+     * @param _userId The user ID of the student to be deleted.
+     * @param jwtToken The JSON Web Token used for authorization.
+     * @return HTTP code 200 with deleted user if successful, else an HTTP 400 or 500 code.
+     */
     @DELETE
     @Path("/{userId}")
-    public String deleteSpecificUser(@PathParam("userId") String _userId){
-        //TODO This method needs to ensure authentication
+    public Response deleteSpecificUser(@PathParam("userId") String _userId, @HeaderParam("jwtToken") String jwtToken){
         try {
-
             int userId = Integer.parseInt(_userId);
             User user = userService.findOne(userId);
-            user = userService.delete(user);
-            String res = jsonb.toJson(user);
-            if(user != null) return res;
-            else return "User ID provided was not formatted properly.";
+
+            JwtConsumer c = new JwtConsumer();
+            String role = a.authUser(c.createJwt(jwtToken));
+
+            if(role.equals("professor")){
+                if(user.getRole().equals("student")){
+                    user = userService.delete(user);
+                    String res = jsonb.toJson(user);
+                    if(user != null)
+                        return new ResponseMessage().sendMessage(res, 200);
+                    else return new ResponseMessage().sendMessage("User not found.", 404);
+                } else return new ResponseMessage().sendMessage(forbid, 403);
+            } else if(role.equals("student"))
+                return new ResponseMessage().sendMessage(forbid, 403);
+            else
+                return new ResponseMessage().sendMessage(unauth, 401);
         } catch (NumberFormatException ne){
-            System.out.println("User ID provided was not formatted properly.");
+            return new ResponseMessage().sendMessage(badnum, 500);
+        } catch (Exception e) {
+            return new ResponseMessage().sendMessage(errorm+e, 500);
         }
-        return null;
     }
 }
